+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/TimedIndexService.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.index.v2.metrics
+package com.digitalasset.platform.apiserver
 
 import java.time.Instant
